Added the mezzanine barrel option. To create a "mezzanine barrel" just add the minimumZ parameter to the barrel definition. This parameter sets the Z for which the mezzanine barrel starts.
This is not optimal for barrels with #modules > 1 (even if it's near to be optimal), because the 
compactification of the barrel is done after modules are placed.

If needed in future it can be added the option of giving maximumZ and nModules and no minimumZ and 
place the modules in a reversed way.